requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-docx==1.1.0
google-generativeai
gunicorn